Apply cargo fix cleanups
This applies current suggestions from `cargo clippy --fix -D warnings`
to ensure that the code is a bit more rust idiomatic.

+++ src/smalltalk/execution.rs
@@ -16,7 +16,7 @@ impl SmalltalkScriptsToExecute {
     }
 
     pub async fn execute(&self, evaluator: &SmalltalkEvaluator<'_, '_>) -> Result<()> {
-        let mut index = 0 as usize;
+        let mut index = 0_usize;
         let total = self.scripts.len();
 
         for script in &self.scripts {
